requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.1.1
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.104.0
greenlet==3.0.1
h11==0.14.0
idna==3.6
install==1.3.5
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.0
pydantic_core==2.14.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
